refactor: move types to correct places
