+++ Applications/AdminApi/src/AdminApi/Controllers/ClientsController.cs
@@ -1,4 +1,4 @@
-using Backbone.AdminApi.Infrastructure.DTOs;
+using Backbone.AdminApi.DTOs;
 using Backbone.AdminApi.Infrastructure.Persistence.Database;
 using Backbone.BuildingBlocks.API;
 using Backbone.BuildingBlocks.API.Mvc;
@@ -28,11 +28,11 @@ public class ClientsController : ApiControllerBase
     }
 
     [HttpGet]
-    [ProducesResponseType(typeof(HttpResponseEnvelopeResult<ClientOverview>), StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(HttpResponseEnvelopeResult<ClientOverviewDTO>), StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<IActionResult> GetAllClients(CancellationToken cancellationToken)
     {
-        var clientOverviews = await _adminApiDbContext.OpenIddictApplications.Select(a => new ClientOverview
+        var clientOverviews = await _adminApiDbContext.OpenIddictApplications.Select(a => new ClientOverviewDTO
         {
             CreatedAt = a.CreatedAt,
             ClientId = a.ClientId,
