+++ Applications/AdminCli/src/AdminCli/Commands/Database/Model/FileExport.cs
@@ -1,13 +1,13 @@
-﻿namespace Backbone.AdminApi.Infrastructure.Persistence.Models.Exports;
+﻿namespace Backbone.AdminCli.Commands.Database.Model;
 
 public class FileExport
 {
     public required string FileId { get; set; }
     public required string CreatedBy { get; set; }
-    public string? CreatedByClientName { get; set; }
+    public required string? CreatedByClientName { get; set; }
     public required DateTime CreatedAt { get; set; }
     public required string Owner { get; set; }
-    public string? OwnerClientName { get; set; }
+    public required string? OwnerClientName { get; set; }
     public required long CipherSize { get; set; }
     public required DateTime ExpiresAt { get; set; }
 }
