+++ Applications/AdminCli/src/AdminCli/Commands/Database/Model/RelationshipTemplateExport.cs
@@ -1,10 +1,10 @@
-﻿namespace Backbone.AdminApi.Infrastructure.Persistence.Models.Exports;
+﻿namespace Backbone.AdminCli.Commands.Database.Model;
 
 public class RelationshipTemplateExport
 {
     public required string TemplateId { get; set; }
     public required string CreatedBy { get; set; }
-    public string? CreatedByClientName { get; set; }
+    public required string? CreatedByClientName { get; set; }
     public required DateTime CreatedAt { get; set; }
     public required DateTime? AllocatedAt { get; set; }
 }
