+++ Applications/AdminCli/src/AdminCli/Commands/Database/Model/SyncErrorExport.cs
@@ -1,10 +1,10 @@
-﻿namespace Backbone.AdminApi.Infrastructure.Persistence.Models.Exports;
+﻿namespace Backbone.AdminCli.Commands.Database.Model;
 
 public class SyncErrorExport
 {
     public required string ErrorId { get; set; }
     public required string SyncItemOwner { get; set; }
-    public string? SyncItemOwnerClientName { get; set; }
+    public required string? SyncItemOwnerClientName { get; set; }
     public required DateTime? CreatedAt { get; set; }
     public required string ErrorCode { get; set; }
 }
